tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.arc.end.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.arc.end</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.arc.end</h1>
<p class="desc">arc() adds the end point of the arc to the subpath</p>


<script>
var t = async_test("arc() adds the end point of the arc to the subpath");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.lineWidth = 50;
  ctx.strokeStyle = '#0f0';
  ctx.beginPath();
  ctx.moveTo(-100, 0);
  ctx.arc(-100, 0, 25, -Math.PI/2, Math.PI/2, true);
  ctx.lineTo(100, 25);
  ctx.stroke();
  _assertPixel(canvas, 50,25, 0,255,0,255);
  t.done();

});
</script>
